Part 6 of RFC2906 - Switch the inheritance source from `workspace` to `workspace.package`
