+++ src/cargo/core/workspace.rs
@@ -23,9 +23,7 @@ use crate::sources::{PathSource, CRATES_IO_INDEX, CRATES_IO_REGISTRY};
 use crate::util::errors::{CargoResult, ManifestError};
 use crate::util::interning::InternedString;
 use crate::util::lev_distance;
-use crate::util::toml::{
-    read_manifest, readme_for_project, StringOrBool, TomlDependency, TomlProfiles, VecStringOrBool,
-};
+use crate::util::toml::{read_manifest, InheritableFields, TomlDependency, TomlProfiles};
 use crate::util::{config::ConfigRelativePath, Config, Filesystem, IntoUrl};
 use cargo_util::paths;
 use cargo_util::paths::normalize_path;
@@ -1641,213 +1639,6 @@ impl WorkspaceRootConfig {
 
     pub fn inheritable(&self) -> &InheritableFields {
         &self.inheritable_fields
-    }
-}
-
-/// A group of fields that are inheritable by members of the workspace
-#[derive(Clone, Debug, Default)]
-pub struct InheritableFields {
-    dependencies: Option<BTreeMap<String, TomlDependency>>,
-    version: Option<semver::Version>,
-    authors: Option<Vec<String>>,
-    description: Option<String>,
-    homepage: Option<String>,
-    documentation: Option<String>,
-    readme: Option<StringOrBool>,
-    keywords: Option<Vec<String>>,
-    categories: Option<Vec<String>>,
-    license: Option<String>,
-    license_file: Option<String>,
-    repository: Option<String>,
-    publish: Option<VecStringOrBool>,
-    edition: Option<String>,
-    badges: Option<BTreeMap<String, BTreeMap<String, String>>>,
-    rust_version: Option<String>,
-    ws_root: PathBuf,
-}
-
-impl InheritableFields {
-    pub fn new(
-        dependencies: Option<BTreeMap<String, TomlDependency>>,
-        version: Option<semver::Version>,
-        authors: Option<Vec<String>>,
-        description: Option<String>,
-        homepage: Option<String>,
-        documentation: Option<String>,
-        readme: Option<StringOrBool>,
-        keywords: Option<Vec<String>>,
-        categories: Option<Vec<String>>,
-        license: Option<String>,
-        license_file: Option<String>,
-        repository: Option<String>,
-        publish: Option<VecStringOrBool>,
-        edition: Option<String>,
-        badges: Option<BTreeMap<String, BTreeMap<String, String>>>,
-        rust_version: Option<String>,
-        ws_root: PathBuf,
-    ) -> InheritableFields {
-        Self {
-            dependencies,
-            version,
-            authors,
-            description,
-            homepage,
-            documentation,
-            readme,
-            keywords,
-            categories,
-            license,
-            license_file,
-            repository,
-            publish,
-            edition,
-            badges,
-            rust_version,
-            ws_root,
-        }
-    }
-
-    pub fn dependencies(&self) -> CargoResult<BTreeMap<String, TomlDependency>> {
-        self.dependencies.clone().map_or(
-            Err(anyhow!("`workspace.dependencies` was not defined")),
-            |d| Ok(d),
-        )
-    }
-
-    pub fn get_dependency(&self, name: &str) -> CargoResult<TomlDependency> {
-        self.dependencies.clone().map_or(
-            Err(anyhow!("`workspace.dependencies` was not defined")),
-            |deps| {
-                deps.get(name).map_or(
-                    Err(anyhow!(
-                        "`dependency.{}` was not found in `workspace.dependencies`",
-                        name
-                    )),
-                    |dep| Ok(dep.clone()),
-                )
-            },
-        )
-    }
-
-    pub fn version(&self) -> CargoResult<semver::Version> {
-        self.version
-            .clone()
-            .map_or(Err(anyhow!("`workspace.version` was not defined")), |d| {
-                Ok(d)
-            })
-    }
-
-    pub fn authors(&self) -> CargoResult<Vec<String>> {
-        self.authors
-            .clone()
-            .map_or(Err(anyhow!("`workspace.authors` was not defined")), |d| {
-                Ok(d)
-            })
-    }
-
-    pub fn description(&self) -> CargoResult<String> {
-        self.description.clone().map_or(
-            Err(anyhow!("`workspace.description` was not defined")),
-            |d| Ok(d),
-        )
-    }
-
-    pub fn homepage(&self) -> CargoResult<String> {
-        self.homepage
-            .clone()
-            .map_or(Err(anyhow!("`workspace.homepage` was not defined")), |d| {
-                Ok(d)
-            })
-    }
-
-    pub fn documentation(&self) -> CargoResult<String> {
-        self.documentation.clone().map_or(
-            Err(anyhow!("`workspace.documentation` was not defined")),
-            |d| Ok(d),
-        )
-    }
-
-    pub fn readme(&self, package_root: &Path) -> CargoResult<StringOrBool> {
-        readme_for_project(self.ws_root.as_path(), self.readme.clone()).map_or(
-            Err(anyhow!("`workspace.readme` was not defined")),
-            |readme| {
-                let rel_path =
-                    resolve_relative_path("readme", &self.ws_root, package_root, &readme)?;
-                Ok(StringOrBool::String(rel_path))
-            },
-        )
-    }
-
-    pub fn keywords(&self) -> CargoResult<Vec<String>> {
-        self.keywords
-            .clone()
-            .map_or(Err(anyhow!("`workspace.keywords` was not defined")), |d| {
-                Ok(d)
-            })
-    }
-
-    pub fn categories(&self) -> CargoResult<Vec<String>> {
-        self.categories.clone().map_or(
-            Err(anyhow!("`workspace.categories` was not defined")),
-            |d| Ok(d),
-        )
-    }
-
-    pub fn license(&self) -> CargoResult<String> {
-        self.license
-            .clone()
-            .map_or(Err(anyhow!("`workspace.license` was not defined")), |d| {
-                Ok(d)
-            })
-    }
-
-    pub fn license_file(&self, package_root: &Path) -> CargoResult<String> {
-        self.license_file.clone().map_or(
-            Err(anyhow!("`workspace.license_file` was not defined")),
-            |d| resolve_relative_path("license-file", &self.ws_root, package_root, &d),
-        )
-    }
-
-    pub fn repository(&self) -> CargoResult<String> {
-        self.repository.clone().map_or(
-            Err(anyhow!("`workspace.repository` was not defined")),
-            |d| Ok(d),
-        )
-    }
-
-    pub fn publish(&self) -> CargoResult<VecStringOrBool> {
-        self.publish
-            .clone()
-            .map_or(Err(anyhow!("`workspace.publish` was not defined")), |d| {
-                Ok(d)
-            })
-    }
-
-    pub fn edition(&self) -> CargoResult<String> {
-        self.edition
-            .clone()
-            .map_or(Err(anyhow!("`workspace.edition` was not defined")), |d| {
-                Ok(d)
-            })
-    }
-
-    pub fn rust_version(&self) -> CargoResult<String> {
-        self.rust_version.clone().map_or(
-            Err(anyhow!("`workspace.rust-version` was not defined")),
-            |d| Ok(d),
-        )
-    }
-
-    pub fn badges(&self) -> CargoResult<BTreeMap<String, BTreeMap<String, String>>> {
-        self.badges
-            .clone()
-            .map_or(Err(anyhow!("`workspace.badges` was not defined")), |d| {
-                Ok(d)
-            })
-    }
-
-    pub fn ws_root(&self) -> &PathBuf {
-        &self.ws_root
     }
 }
